fix: incorrect parameter type

+++ PaintJob/src/Main.java
@@ -50,7 +50,7 @@ public class Main {
      * @return The number of buckets needed to purchase before going to work.
      */
     public static int getBucketCount(double width, double height,
-            double areaPerBucket, double extraBuckets){
+            double areaPerBucket, int extraBuckets){
         
         if(width <= 0 || height <= 0 || areaPerBucket <= 0 || extraBuckets < 0){
             return -1;
